organize project, pacakge by type
