+++ app/src/main/AndroidManifest.xml
@@ -9,8 +9,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ModelManagerActivity" />
-        <activity android:name=".MainActivity">
+        <activity android:name=".activities.ModelManagerActivity" />
+        <activity android:name=".activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
