+++ app/src/main/res/layout/activity_model_manager.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ModelManagerActivity">
+    tools:context=".activities.ModelManagerActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
